+++ CHANGES.rst
@@ -1,4 +1,13 @@
-version 0.1 (2013-06-02)
+version 1.0 (02-Jun-2013)
 ------------------------
 
 - Initial release.
+
+
+version 1.1 (19-Aug-2013)
+------------------------
+
+- Changed the program flow to use database extraction by default. This can provide dramatic 
+  speedups for the use case of calling a large database using a small number of citations,
+  when there are no new citations added to the auxiliary file.
+